- Various packet clarifications and using new typed skip and unk functionality  - Thanks ED for information related to the ban times in the login packet  - Thanks Fraysa for idears
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3472 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/LoginServer/LoginPacket.hpp
@@ -53,7 +53,7 @@ namespace LoginPacket {
 	}
 
 	PACKET(loginError, int16_t errorId);
-	PACKET(loginBan, int8_t reason, int32_t expire);
+	PACKET(loginBan, int8_t reason, int64_t expire);
 	PACKET(loginProcess, int8_t id);
 	PACKET(loginConnect, UserConnection *user, const string_t &username);
 	PACKET(pinAssigned);
